requirements.txt
pandas
numpy
scikit-learn
joblib
mysql-connector-python
